combined codemirror component with quizcode component, trying to implement the time
